requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dataclasses-json==0.5.7
discord.py @ git+https://github.com/Rapptz/discord.py@b3a24846bddc894ae6d9bfb49ed8ee93b2d12fcb
dnspython==2.2.1
frozenlist==1.3.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
isodate==0.6.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.7.2
